requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
pandas
numpy
matplotlib
scipy
langchain-google-genai
Pillow